arc: Add support to cache extracted raw videos in encode tests.
As the raw video files used to test encoding are too large to download
and store on a DUT, videos are stored in the vp9.webm format. Before
each test these videos are extracted to create the raw input video.
This process can take a lot of time, so this CL adds support to cache
the extracted raw video file. This saves about 30s test time when
running 1080p encode testcases. The feature is disabled by default as
we don't want to have DUT's run out of disk space in the lab.

The option is added to both arc++/arcvm variants of these tests:
- arc.VideoEncodeAccel.*
- arc.VideoEncodeAccelPerf.*
- arc.PowerVideoEncodePerf.*

BUG=b:169725788
TEST=various video encode tests on Hatch

+++ src/chromiumos/tast/local/bundles/cros/arc/power_video_encode_perf.go
@@ -25,11 +25,13 @@ import (
 )
 
 const (
-	pvepIterationCount    = 30
-	pvepIterationDuration = 10 * time.Second
-	pvepWarmupDuration    = 10 * time.Second
-	pvepTestSlack         = 5 * time.Minute
-	pvepPowerTestDuration = pvepIterationCount*pvepIterationDuration + pvepWarmupDuration + pvepTestSlack
+	// Enable to cache the extracted raw video to speed up the test.
+	pvepCacheExtractedVideo = false
+	pvepIterationCount      = 30
+	pvepIterationDuration   = 10 * time.Second
+	pvepWarmupDuration      = 10 * time.Second
+	pvepTestSlack           = 5 * time.Minute
+	pvepPowerTestDuration   = pvepIterationCount*pvepIterationDuration + pvepWarmupDuration + pvepTestSlack
 )
 
 func init() {
@@ -98,7 +100,9 @@ func PowerVideoEncodePerf(ctx context.Context, s *testing.State) {
 	if err != nil {
 		s.Fatal("Failed to prepare YUV file: ", err)
 	}
-	defer os.Remove(rawVideoPath)
+	if !pvepCacheExtractedVideo {
+		defer os.Remove(rawVideoPath)
+	}
 
 	// Set up device for measuring power drain.
 	sup, cleanup := setup.New("video power")
